requirements.txt
anyio==3.6.2
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
contextlib2==21.6.0
contextvars==2.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
easyocr==1.6.2
fastapi==0.83.0
h11==0.13.0
idna==3.4
imageio==2.15.0
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
ninja==1.11.1
numpy==1.19.5
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.1.5
pep517==0.13.0
Pillow==8.4.0
pip-tools==6.4.0
pyclipper==1.3.0.post4
pydantic==1.9.2
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.7
python-multipart==0.0.5
pytz==2022.6
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
scikit-image==0.17.2
scipy==1.5.4
seaborn==0.11.2
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tifffile==2020.9.3
tomli==1.2.3
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.12
uvicorn==0.16.0
zipp==3.6.0
